Add describe SOLID (O) violations
