+++ bin/views/fxml/media_cd.fxml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.TitledPane?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.RowConstraints?>
+
+<TitledPane animated="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="320.0" prefWidth="700.0" text="Create new CD Media Item:" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/21">
+  <content>
+    <AnchorPane prefHeight="316.0" prefWidth="598.0">
+         <children>
+            <GridPane layoutX="14.0" layoutY="14.0" prefHeight="226.0" prefWidth="382.0">
+              <columnConstraints>
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+              </columnConstraints>
+              <rowConstraints>
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+              </rowConstraints>
+               <children>
+                  <Label text="Label" />
+                  <Label text="Label" GridPane.rowIndex="1" />
+                  <Label text="Label" GridPane.rowIndex="2" />
+                  <Label text="Label" GridPane.rowIndex="3" />
+                  <ComboBox prefHeight="26.0" prefWidth="194.0" GridPane.columnIndex="1" />
+                  <TextField GridPane.columnIndex="1" GridPane.rowIndex="1" />
+                  <TextField GridPane.columnIndex="1" GridPane.rowIndex="2" />
+                  <DatePicker GridPane.columnIndex="1" GridPane.rowIndex="3" />
+               </children>
+            </GridPane>
+            <Button layoutX="603.0" layoutY="240.0" mnemonicParsing="false" prefHeight="34.0" prefWidth="86.0" text="Create" />
+            <ImageView fitHeight="150.0" fitWidth="236.0" layoutX="440.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true" />
+            <Label layoutX="440.0" layoutY="172.0" prefHeight="18.0" prefWidth="206.0" text="Choose the image for this media" />
+            <ComboBox layoutX="440.0" layoutY="196.0" prefHeight="28.0" prefWidth="236.0" promptText="Image URL..." />
+         </children></AnchorPane>
+  </content>
+</TitledPane>
